refactor(core): move `RouteConfig` to `fsRoutes()` source
