+++ src/mod.ts
@@ -1,6 +1,10 @@
 export { App, type ListenOptions } from "./app.ts";
 export { trailingSlashes } from "./middlewares/trailing_slashes.ts";
-export { fsRoutes, type FsRoutesOptions } from "./plugins/fs_routes/mod.ts";
+export {
+  fsRoutes,
+  type FsRoutesOptions,
+  type RouteConfig,
+} from "./plugins/fs_routes/mod.ts";
 export {
   type HandlerByMethod,
   type HandlerFn,
@@ -9,7 +13,7 @@ export {
   type RouteData,
   type RouteHandler,
 } from "./handlers.ts";
-export type { LayoutConfig, RouteConfig } from "./types.ts";
+export type { LayoutConfig } from "./types.ts";
 export type { Middleware, MiddlewareFn } from "./middlewares/mod.ts";
 export { staticFiles } from "./middlewares/static_files.ts";
 export { cors, type CORSOptions } from "./middlewares/cors.ts";
