+++ src/plugins/fs_routes/mod.ts
@@ -2,7 +2,6 @@ import type { AnyComponent } from "preact";
 import type { App } from "../../app.ts";
 import type { WalkEntry } from "@std/fs/walk";
 import * as path from "@std/path";
-import type { RouteConfig } from "../../types.ts";
 import type { RouteHandler } from "../../handlers.ts";
 import type { MiddlewareFn } from "../../middlewares/mod.ts";
 import {
@@ -29,6 +28,36 @@ interface InternalRoute<State> {
     | RouteHandler<unknown, State>
     | null;
   component: AnyComponent<PageProps<unknown, State>> | null;
+}
+
+export interface RouteConfig {
+  /**
+   * A route override for the page. This is useful for pages where the route
+   * can not be expressed through the filesystem routing capabilities.
+   *
+   * The route override must be a path-to-regexp compatible route matcher.
+   */
+  routeOverride?: string;
+
+  /**
+   * If Content-Security-Policy should be enabled for this page. If 'true', a
+   * locked down policy will be used that allows only the scripts and styles
+   * that are generated by Fresh. Additional scripts and styles can be added
+   * using the `useCSP` hook.
+   */
+  csp?: boolean;
+
+  /**
+   * Skip already inherited layouts
+   * Default: `false`
+   */
+  skipInheritedLayouts?: boolean;
+
+  /**
+   * Skip rendering the `routes/_app` template
+   * Default: `false`
+   */
+  skipAppWrapper?: boolean;
 }
 
 export interface FreshFsItem<State> {
