+++ src/types.ts
@@ -1,33 +1,3 @@
-export interface RouteConfig {
-  /**
-   * A route override for the page. This is useful for pages where the route
-   * can not be expressed through the filesystem routing capabilities.
-   *
-   * The route override must be a path-to-regexp compatible route matcher.
-   */
-  routeOverride?: string;
-
-  /**
-   * If Content-Security-Policy should be enabled for this page. If 'true', a
-   * locked down policy will be used that allows only the scripts and styles
-   * that are generated by Fresh. Additional scripts and styles can be added
-   * using the `useCSP` hook.
-   */
-  csp?: boolean;
-
-  /**
-   * Skip already inherited layouts
-   * Default: `false`
-   */
-  skipInheritedLayouts?: boolean;
-
-  /**
-   * Skip rendering the `routes/_app` template
-   * Default: `false`
-   */
-  skipAppWrapper?: boolean;
-}
-
 export interface LayoutConfig {
   /**
    * Skip already inherited layouts
